Update
Updated weather values to never have rain, or fog.

Provided comments for all created functions for a better understanding of what I  am doing in each.

Changed addMag to take a TStringArray for all mags instead of seperate strings for different mags. Also, mags will be added to your hot bar, 3 and 4 for primary, and 5 and 6 for secondary. Please make sure that you have enough slots in your gear to take those mags.

Removal of many EntityAI objects, didn't need to store them unless I needed to access them later.

Creation of battery check, the TStringArray for batteryCompatibility is incomplete, and could possible have more attacchments. More will be added as they are brought to my attention, or they are released.

Creation of arm_player_secondary, this allows the pistol/secondary to be added to the inventory rather than adding it to the hands.

Return type for arm_player/arm_player_secondary changed from void to EntityAI, so I could add the weapons to QuickBar later.

Added QuickBar settings for primary, and secondary.